accuro_adapter/requirements.txt
httpx[http2]
python-dotenv
pydantic
fastapi
uvicorn[standard]